+++ export/basic_items/data/definition.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="iso-8859-1"?>
+<root version="2.2">
+	<name>Basic Items</name>
+	<author>skelekon</author>
+	<ruleset>4E</ruleset>
+</root>
